requirements.txt
chromadb==0.5.11
datasets==3.0.1
lancedb==0.14.0
langchain==0.3.3
langchain_community==0.3.2
langchain_core==0.3.10
langchain_openai==0.2.2
nest_asyncio==1.6.0
openai==1.51.2
pandas==2.2.3
Pillow==10.4.0
PyPDF2==3.0.1
python-dotenv==1.0.1
ragas==0.1.20
tqdm==4.66.5
utils==1.0.2
